Interim commit
Small fixes, added logic for spawning players into the game when they join/host sessions and created new map (and gamemode) to accommodate these updates.

TODO: add an in-game menu to leave game etc., simple UI and find a way to show mouse on screen without messing with the inputs.